+++ src/main/java/np/edu/nast/payroll/Payroll/repository/AttendanceRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface AttendanceRepository extends JpaRepository<Attendance, Long> {
+public interface AttendanceRepository extends JpaRepository<Attendance, Integer> {
 
-    List<Attendance> findByEmployeeEmpId(Long empId);
+    List<Attendance> findByEmployeeEmpId(Integer empId);
 }
